revised create account screen
